Add frames to button inputs operator
